Move per-cell nudge operations off the cell class
Turn the four cell nudge methods into free functions in nudge_cell_ops that
take the cell as an argument, and leave thin delegators behind so existing
call sites and tests are unchanged.

Nudging needs to build a throwaway cell when it confirms a positive-Z contact,
so nudge code has to sit above the cell class in the import order. Methods on
cell would force the opposite. Rename _clear_geometry to clear_geometry now
that it is called from outside the class.

Emitted meshes are unchanged: all 66 meshes across all 48 launch configs stay
byte identical.
